+++ packages/suite/src/components/suite/SwitchDevice/components/DeviceItem/components/ColHeader/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { Icon, Tooltip, TooltipProps, useTheme, variables } from '@trezor/components';
+import { Tooltip, TooltipProps, variables } from '@trezor/components';
 
 const Wrapper = styled.div`
     display: flex;
@@ -22,19 +22,14 @@ interface Props {
 }
 
 const ColHeader = ({ children, tooltipContent, ...rest }: Props) => {
-    const theme = useTheme();
     return (
         <Wrapper {...rest}>
-            <Text>{children}</Text>
-            {tooltipContent && (
-                <Tooltip maxWidth={285} placement="top" content={tooltipContent}>
-                    <Icon
-                        icon="INFO_ACTIVE"
-                        color={theme.TYPE_LIGHT_GREY}
-                        size={16}
-                        useCursorPointer
-                    />
+            {tooltipContent ? (
+                <Tooltip maxWidth={285} content={tooltipContent} dashed>
+                    <Text>{children}</Text>
                 </Tooltip>
+            ) : (
+                <Text>{children}</Text>
             )}
         </Wrapper>
     );